Move unmarshal into server code and add Link to MsgTraceEgress
Moving the unmarshal to server code helps with the library importing
the server code and be able to reuse that code.
Added a Link field to a MsgTraceEgress that the server does not
do anything with, but library will be able to link trace events
together.
